fix(questions): make Cancel use secondary style and lock hover (#524)
